+++ tests/providers/google/marketing_platform/operators/test_display_video.py
@@ -25,11 +25,14 @@ import pytest
 
 from airflow.models import DAG, TaskInstance as TI
 from airflow.providers.google.marketing_platform.operators.display_video import (
+    GoogleDisplayVideo360CreateQueryOperator,
     GoogleDisplayVideo360CreateReportOperator,
     GoogleDisplayVideo360CreateSDFDownloadTaskOperator,
     GoogleDisplayVideo360DeleteReportOperator,
     GoogleDisplayVideo360DownloadLineItemsOperator,
     GoogleDisplayVideo360DownloadReportOperator,
+    GoogleDisplayVideo360DownloadReportV2Operator,
+    GoogleDisplayVideo360RunQueryOperator,
     GoogleDisplayVideo360RunReportOperator,
     GoogleDisplayVideo360SDFtoGCSOperator,
     GoogleDisplayVideo360UploadLineItemsOperator,
@@ -37,7 +40,7 @@ from airflow.providers.google.marketing_platform.operators.display_video import 
 from airflow.utils import timezone
 from airflow.utils.session import create_session
 
-API_VERSION = "api_version"
+API_VERSION = "v2"
 GCP_CONN_ID = "google_cloud_default"
 DELEGATE_TO: str | None = None
 IMPERSONATION_CHAIN = ["ACCOUNT_1", "ACCOUNT_2", "ACCOUNT_3"]
@@ -235,6 +238,132 @@ class TestGoogleDisplayVideo360DownloadReportOperator:
         )
 
 
+class TestGoogleDisplayVideo360DownloadReportV2Operator:
+    def setup_method(self):
+        with create_session() as session:
+            session.query(TI).delete()
+
+    def teardown_method(self):
+        with create_session() as session:
+            session.query(TI).delete()
+
+    @mock.patch("airflow.providers.google.marketing_platform.operators.display_video.shutil")
+    @mock.patch("airflow.providers.google.marketing_platform.operators.display_video.urllib.request")
+    @mock.patch("airflow.providers.google.marketing_platform.operators.display_video.tempfile")
+    @mock.patch(
+        "airflow.providers.google.marketing_platform.operators."
+        "display_video.GoogleDisplayVideo360DownloadReportV2Operator.xcom_push"
+    )
+    @mock.patch("airflow.providers.google.marketing_platform.operators.display_video.GCSHook")
+    @mock.patch(
+        "airflow.providers.google.marketing_platform.operators.display_video.GoogleDisplayVideo360Hook"
+    )
+    def test_execute(
+        self,
+        mock_hook,
+        mock_gcs_hook,
+        mock_xcom,
+        mock_temp,
+        mock_request,
+        mock_shutil,
+    ):
+        mock_temp.NamedTemporaryFile.return_value.__enter__.return_value.name = FILENAME
+        mock_hook.return_value.get_report.return_value = {
+            "metadata": {
+                "status": {"state": "DONE"},
+                "googleCloudStoragePath": "TEST",
+            }
+        }
+        op = GoogleDisplayVideo360DownloadReportV2Operator(
+            query_id=QUERY_ID,
+            report_id=REPORT_ID,
+            bucket_name=BUCKET_NAME,
+            report_name=REPORT_NAME,
+            task_id="test_task",
+        )
+        op.execute(context=None)
+        mock_hook.assert_called_once_with(
+            gcp_conn_id=GCP_CONN_ID,
+            delegate_to=None,
+            api_version="v2",
+            impersonation_chain=None,
+        )
+        mock_hook.return_value.get_report.assert_called_once_with(report_id=REPORT_ID, query_id=QUERY_ID)
+
+        mock_gcs_hook.assert_called_once_with(
+            gcp_conn_id=GCP_CONN_ID,
+            delegate_to=None,
+            impersonation_chain=None,
+        )
+        mock_gcs_hook.return_value.upload.assert_called_once_with(
+            bucket_name=BUCKET_NAME,
+            filename=FILENAME,
+            gzip=True,
+            mime_type="text/csv",
+            object_name=REPORT_NAME + ".gz",
+        )
+        mock_xcom.assert_called_once_with(None, key="report_name", value=REPORT_NAME + ".gz")
+
+    @pytest.mark.parametrize(
+        "test_bucket_name",
+        [BUCKET_NAME, f"gs://{BUCKET_NAME}", "XComArg", "{{ ti.xcom_pull(task_ids='f') }}"],
+    )
+    @mock.patch("airflow.providers.google.marketing_platform.operators.display_video.shutil")
+    @mock.patch("airflow.providers.google.marketing_platform.operators.display_video.urllib.request")
+    @mock.patch("airflow.providers.google.marketing_platform.operators.display_video.tempfile")
+    @mock.patch("airflow.providers.google.marketing_platform.operators.display_video.GCSHook")
+    @mock.patch(
+        "airflow.providers.google.marketing_platform.operators.display_video.GoogleDisplayVideo360Hook"
+    )
+    def test_set_bucket_name(
+        self,
+        mock_hook,
+        mock_gcs_hook,
+        mock_temp,
+        mock_request,
+        mock_shutil,
+        test_bucket_name,
+    ):
+        mock_temp.NamedTemporaryFile.return_value.__enter__.return_value.name = FILENAME
+        mock_hook.return_value.get_report.return_value = {
+            "metadata": {"status": {"state": "DONE"}, "googleCloudStoragePath": "TEST"}
+        }
+
+        dag = DAG(
+            dag_id="test_set_bucket_name",
+            start_date=DEFAULT_DATE,
+            schedule=None,
+            catchup=False,
+        )
+
+        if BUCKET_NAME not in test_bucket_name:
+
+            @dag.task
+            def f():
+                return BUCKET_NAME
+
+            taskflow_op = f()
+            taskflow_op.operator.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE)
+
+        op = GoogleDisplayVideo360DownloadReportV2Operator(
+            query_id=QUERY_ID,
+            report_id=REPORT_ID,
+            bucket_name=test_bucket_name if test_bucket_name != "XComArg" else taskflow_op,
+            report_name=REPORT_NAME,
+            task_id="test_task",
+            dag=dag,
+        )
+        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE)
+
+        mock_gcs_hook.return_value.upload.assert_called_once_with(
+            bucket_name=BUCKET_NAME,
+            filename=FILENAME,
+            gzip=True,
+            mime_type="text/csv",
+            object_name=REPORT_NAME + ".gz",
+        )
+
+
 class TestGoogleDisplayVideo360RunReportOperator:
     @mock.patch(
         "airflow.providers.google.marketing_platform.operators.display_video.GoogleDisplayVideo360Hook"
@@ -255,6 +384,41 @@ class TestGoogleDisplayVideo360RunReportOperator:
             impersonation_chain=None,
         )
         hook_mock.return_value.run_query.assert_called_once_with(query_id=REPORT_ID, params=parameters)
+
+
+class TestGoogleDisplayVideo360RunQueryOperator:
+    @mock.patch(
+        "airflow.providers.google.marketing_platform.operators."
+        "display_video.GoogleDisplayVideo360RunQueryOperator.xcom_push"
+    )
+    @mock.patch(
+        "airflow.providers.google.marketing_platform.operators.display_video.GoogleDisplayVideo360Hook"
+    )
+    def test_execute(self, hook_mock, mock_xcom):
+        parameters = {"param": "test"}
+        hook_mock.return_value.run_query.return_value = {
+            "key": {
+                "queryId": QUERY_ID,
+                "reportId": REPORT_ID,
+            }
+        }
+        op = GoogleDisplayVideo360RunQueryOperator(
+            query_id=QUERY_ID,
+            parameters=parameters,
+            api_version=API_VERSION,
+            task_id="test_task",
+        )
+        op.execute(context=None)
+        hook_mock.assert_called_once_with(
+            gcp_conn_id=GCP_CONN_ID,
+            delegate_to=None,
+            api_version=API_VERSION,
+            impersonation_chain=None,
+        )
+
+        mock_xcom.assert_any_call(None, key="query_id", value=QUERY_ID)
+        mock_xcom.assert_any_call(None, key="report_id", value=REPORT_ID)
+        hook_mock.return_value.run_query.assert_called_once_with(query_id=QUERY_ID, params=parameters)
 
 
 class TestGoogleDisplayVideo360DownloadLineItemsOperator:
@@ -456,3 +620,37 @@ class TestGoogleDisplayVideo360CreateSDFDownloadTaskOperator:
             body_request=body_request
         )
         xcom_mock.assert_called_once_with(None, key="name", value=test_name)
+
+
+class TestGoogleDisplayVideo360CreateQueryOperator:
+    @mock.patch(
+        "airflow.providers.google.marketing_platform.operators."
+        "display_video.GoogleDisplayVideo360CreateQueryOperator.xcom_push"
+    )
+    @mock.patch(
+        "airflow.providers.google.marketing_platform.operators.display_video.GoogleDisplayVideo360Hook"
+    )
+    def test_execute(self, hook_mock, xcom_mock):
+        body = {"body": "test"}
+        hook_mock.return_value.create_query.return_value = {"queryId": QUERY_ID}
+        op = GoogleDisplayVideo360CreateQueryOperator(body=body, task_id="test_task")
+        op.execute(context=None)
+        hook_mock.assert_called_once_with(
+            gcp_conn_id=GCP_CONN_ID,
+            delegate_to=None,
+            api_version="v2",
+            impersonation_chain=None,
+        )
+        hook_mock.return_value.create_query.assert_called_once_with(query=body)
+        xcom_mock.assert_called_once_with(None, key="query_id", value=QUERY_ID)
+
+    def test_prepare_template(self):
+        body = {"key": "value"}
+        with NamedTemporaryFile("w+", suffix=".json") as f:
+            f.write(json.dumps(body))
+            f.flush()
+            op = GoogleDisplayVideo360CreateQueryOperator(body=body, task_id="test_task")
+            op.prepare_template()
+
+        assert isinstance(op.body, dict)
+        assert op.body == body