+++ airflow/providers/google/CHANGELOG.rst
@@ -23,8 +23,14 @@
 Changelog
 ---------
 
-8.12.1
-......
+9.0.0
+.....
+
+Breaking changes
+~~~~~~~~~~~~~~~~
+
+Google  announced sunset of Bid manager API v1 and v1.1 by April 27, 2023 for more information
+please check: `docs <https://developers.google.com/bid-manager/v1.1>`_  As a result default value of api_version in GoogleDisplayVideo360Hook and related operators updated to v2
 
 This version of provider contains a temporary workaround to issue with ``v11`` version of
 google-ads API being discontinued, while the google provider dependencies preventing installing